@@ -1,68 +1,69 @@
 {
-  "name": "cloud-storage-app",
-  "version": "1.0.1",
-  "description": "API Frontend for the ARK Cloud Storage Mod",
-  "homepage": "https://www.curseforge.com/ark-survival-ascended/mods/cloud-storage",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "start:dev": "ng serve --watch --configuration development",
-    "build": "ng build",
-    "build:dev": "ng build --watch --configuration development",
-    "lint": "ng lint",
-    "test": "ng test",
-    "deploy": "semantic-release",
-    "prepare": "husky"
-  },
-  "private": false,
-  "license": "MIT",
-  "dependencies": {
-    "@angular/animations": "~20.3.6",
-    "@angular/cdk": "~20.2.9",
-    "@angular/common": "~20.3.6",
-    "@angular/compiler": "~20.3.6",
-    "@angular/core": "~20.3.6",
-    "@angular/forms": "~20.3.6",
-    "@angular/material": "~20.2.9",
-    "@angular/platform-browser": "~20.3.6",
-    "@angular/platform-browser-dynamic": "~20.3.6",
-    "@angular/router": "~20.3.6",
-    "rxjs": "~7.8.0",
-    "tslib": "~2.3.0",
-    "zone.js": "~0.15.1"
-  },
-  "devDependencies": {
-    "@angular-eslint/builder": "~18.3.1",
-    "@angular-eslint/eslint-plugin": "~18.3.1",
-    "@angular-eslint/eslint-plugin-template": "~18.3.1",
-    "@angular-eslint/schematics": "~18.3.1",
-    "@angular-eslint/template-parser": "~18.3.1",
-    "@angular/build": "^20.3.6",
-    "@angular/cli": "~20.3.6",
-    "@angular/compiler-cli": "~20.3.6",
-    "@codedependant/semantic-release-docker": "~5.0.3",
-    "@commitlint/cli": "~19.5.0",
-    "@commitlint/config-conventional": "~19.5.0",
-    "@semantic-release/git": "~10.0.1",
-    "@semantic-release/github": "~11.0.0",
-    "@semantic-release/npm": "~12.0.1",
-    "@types/jasmine": "~5.1.0",
-    "@typescript-eslint/eslint-plugin": "~7.11.0",
-    "@typescript-eslint/parser": "~7.11.0",
-    "commitlint": "~19.5.0",
-    "eslint": "~8.57.0",
-    "eslint-config-prettier": "~9.1.0",
-    "eslint-plugin-prettier": "~5.2.1",
-    "husky": "~9.1.6",
-    "jasmine-core": "~5.2.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.1.0",
-    "lint-staged": "~15.2.10",
-    "prettier": "~3.3.3",
-    "semantic-release": "~24.1.3",
-    "typescript": "~5.9.3"
-  }
+    "name": "cloud-storage-app",
+    "version": "1.0.1",
+    "description": "API Frontend for the ARK Cloud Storage Mod",
+    "homepage": "https://www.curseforge.com/ark-survival-ascended/mods/cloud-storage",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "start:dev": "ng serve --watch --configuration development",
+        "build": "ng build",
+        "build:dev": "ng build --watch --configuration development",
+        "lint": "ng lint",
+        "test": "ng test",
+        "deploy": "semantic-release",
+        "prepare": "husky"
+    },
+    "private": false,
+    "license": "MIT",
+    "dependencies": {
+        "@angular/animations": "~20.3.6",
+        "@angular/cdk": "~20.2.9",
+        "@angular/common": "~20.3.6",
+        "@angular/compiler": "~20.3.6",
+        "@angular/core": "~20.3.6",
+        "@angular/forms": "~20.3.6",
+        "@angular/material": "~20.2.9",
+        "@angular/platform-browser": "~20.3.6",
+        "@angular/platform-browser-dynamic": "~20.3.6",
+        "@angular/router": "~20.3.6",
+        "rxjs": "~7.8.2",
+        "tslib": "~2.8.1",
+        "zone.js": "~0.15.1"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~20.3.6",
+        "@angular-eslint/builder": "~20.4.0",
+        "@angular-eslint/eslint-plugin": "~20.4.0",
+        "@angular-eslint/eslint-plugin-template": "~20.4.0",
+        "@angular-eslint/schematics": "~20.4.0",
+        "@angular-eslint/template-parser": "~20.4.0",
+        "@angular/build": "^20.3.6",
+        "@angular/cli": "~20.3.6",
+        "@angular/compiler-cli": "~20.3.6",
+        "@codedependant/semantic-release-docker": "~5.1.1",
+        "@commitlint/cli": "~20.1.0",
+        "@commitlint/config-conventional": "~20.0.0",
+        "@semantic-release/git": "~10.0.1",
+        "@semantic-release/github": "~12.0.0",
+        "@semantic-release/npm": "~13.0.0",
+        "@types/jasmine": "~5.1.12",
+        "@typescript-eslint/eslint-plugin": "~8.46.1",
+        "@typescript-eslint/parser": "~8.46.1",
+        "commitlint": "~20.1.0",
+        "eslint": "~9.38.0",
+        "eslint-config-prettier": "~10.1.8",
+        "eslint-plugin-prettier": "~5.5.4",
+        "husky": "~9.1.7",
+        "jasmine-core": "~5.12.0",
+        "karma": "~6.4.4",
+        "karma-chrome-launcher": "~3.2.0",
+        "karma-coverage": "~2.2.1",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.1.0",
+        "lint-staged": "~16.2.4",
+        "prettier": "~3.6.2",
+        "semantic-release": "~25.0.0",
+        "typescript": "~5.9.3"
+    }
 }
