@@ -16,7 +16,7 @@ import { environment } from "../environments/environment";
         MatButton,
     ],
     templateUrl: "./app.component.html",
-    styleUrl: "./app.component.scss"
+    styleUrl: "./app.component.scss",
 })
 export class AppComponent {
     protected openDocs(): void {
