@@ -5,7 +5,7 @@ import { environment } from "../../environments/environment";
     selector: "app-home",
     imports: [],
     templateUrl: "./home.component.html",
-    styleUrl: "./home.component.scss"
+    styleUrl: "./home.component.scss",
 })
 export class HomeComponent {
     protected documentation = environment.documentation;
